requirements.txt
Pillow==9.4.0
pyinstaller==5.7
pyinstaller-hooks-contrib==2022.8
PySide6-Essentials==6.5.0
